+++ firmware/controllers/can/can_dash.h
@@ -1,0 +1,9 @@
+// todo: header
+
+
+#pragma once
+
+void canDashboardBMW();
+void canDashboardFiat();
+void canDashboardVAG();
+void canMazdaRX8();
