+++ firmware/controllers/controllers.mk
@@ -36,9 +36,11 @@ CONTROLLERS_SRC_CPP = \
 	$(CONTROLLERS_DIR)/injector_central.cpp \
 	$(CONTROLLERS_DIR)/can/obd2.cpp \
 	$(CONTROLLERS_DIR)/can/can_rx.cpp \
+	$(CONTROLLERS_DIR)/can/can_tx.cpp \
+	$(CONTROLLERS_DIR)/can/can_dash.cpp \
  	$(CONTROLLERS_DIR)/engine_controller.cpp \
 	$(CONTROLLERS_DIR)/persistent_store.cpp \
-	
+
 
 CONTROLLERS_INC=\
 	$(CONTROLLERS_DIR) \
